Update class, insert SQL file

+++ src/main/java/me/tadasdev/screens/usersScreen/AuthorScreen.java
@@ -3,6 +3,7 @@ package me.tadasdev.screens.usersScreen;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
@@ -21,29 +22,40 @@ import javafx.stage.Stage;
 import javafx.util.Callback;
 import me.tadasdev.repositories.AuthorRepository;
 import me.tadasdev.screens.MenuScreen;
+import me.tadasdev.screens.refactorScreens.UpdateLibrary;
 import me.tadasdev.users.Author;
 
 import javax.persistence.Table;
+import javax.persistence.criteria.Selection;
 
 
 public class AuthorScreen {
     ObservableList<Author> authorsList = FXCollections.observableArrayList(AuthorRepository.getAllList());
 
+
     TableView<Author> authorTable = new TableView<>();
+    TableView<Author> authorTableView = authorTable;
     public String stackName = "First Name";
 
-    public void authorsScreen(Stage stage){
+    public void authorsScreen(Stage stage, boolean deleteButtonBool, boolean updateButtonBool){
 
-        TableView tableView = this.tableView();
+        TableView tableView = this.tableView(deleteButtonBool, updateButtonBool);
 
         Button backButton = new Button("Back");
-        backButton.setOnAction(event -> MenuScreen.menuScreen(stage));
+        backButton.setOnAction(event -> {
+            if(deleteButtonBool || updateButtonBool){
+                UpdateLibrary updateLibrary = new UpdateLibrary();
+                updateLibrary.updateLibrary(stage);
+            }else{
+                MenuScreen.menuScreen(stage);
+            }
+        });
         Button addButton = new Button("add");
         TextField firstNamee = new TextField();
         TextField lastNamee = new TextField();
 
         addButton.setOnAction(event -> {
-            AuthorRepository.save(firstNamee.getText(), lastNamee.getText());
+            AuthorRepository.addAuthor(firstNamee.getText(), lastNamee.getText());
             firstNamee.setText("");
             lastNamee.setText("");
             ObservableList<Author> authorList = FXCollections.observableArrayList(AuthorRepository.getAllList());
@@ -52,14 +64,14 @@ public class AuthorScreen {
         HBox hBox = new HBox();
         hBox.getChildren().addAll(firstNamee, lastNamee, addButton);
 
-        VBox vBox = new VBox();
-        vBox.getChildren().addAll(backButton);
-        vBox.getChildren().add(addSearchBar(stage));
-        vBox.getChildren().add(tableView);
-        vBox.getChildren().add(hBox);
-        vBox.setMinSize(500,500);
+        VBox root = new VBox();
+        root.getChildren().addAll(backButton);
+        root.getChildren().add(addSearchBar(stage, deleteButtonBool, updateButtonBool));
+        root.getChildren().add(tableView);
+        root.getChildren().add(hBox);
+        root.setMinSize(500,500);
 
-        Scene scene = new Scene(vBox);
+        Scene scene = new Scene(root);
         stage.setScene(scene);
         stage.setMinWidth(500);
         stage.setMinHeight(500);
@@ -67,7 +79,7 @@ public class AuthorScreen {
 
     }
 
-    public TableView tableView(){
+    public TableView tableView(boolean deleteButtonBool, boolean updateButtonBool){
         ObservableList<Author> authorsList = this.authorsList;
 
         TableColumn<Author, Integer> id = new TableColumn<>("Id");
@@ -80,23 +92,50 @@ public class AuthorScreen {
         firstName.setCellValueFactory(new PropertyValueFactory<>("first_name"));
         firstName.setCellFactory(TextFieldTableCell.forTableColumn());
         firstName.setMinWidth(80);
+        firstName.setOnEditCommit(
+                new EventHandler<TableColumn.CellEditEvent<Author, String>>() {
+                    @Override
+                    public void handle(TableColumn.CellEditEvent<Author, String> t) {
+                        ((Author) t.getTableView().getItems().get(
+                                t.getTablePosition().getRow())
+                        ).setFirst_name(t.getNewValue());
+                    }
+                }
+        );
 
         lastName.setCellValueFactory(new PropertyValueFactory<>("last_name"));
         lastName.setCellFactory(TextFieldTableCell.forTableColumn());
         lastName.setMinWidth(80);
+        lastName.setOnEditCommit(
+                new EventHandler<TableColumn.CellEditEvent<Author, String>>() {
+                    @Override
+                    public void handle(TableColumn.CellEditEvent<Author, String> event) {
+                        ((Author) event.getTableView().getItems().get(
+                                event.getTablePosition().getRow())
+                        ).setLast_name(event.getNewValue());
+                    }
+                }
+        );
 
         authorTable.getColumns().setAll(id, firstName, lastName);
         //authorTable.prefWidth(20);
         authorTable.setItems(authorsList);
-        deleteButton();
+        if(deleteButtonBool){
+            deleteButton();
+        }
+        if(updateButtonBool){
+            authorTable.setEditable(true);
+            updateButton();
+
+        }else{
+            authorTable.setEditable(false);
+        }
 
         return authorTable;
     }
 
-
-
     public void deleteButton(){
-        TableColumn<Author, Void> buttons = new TableColumn("Action");
+        TableColumn<Author, Void> buttons = new TableColumn("Delete");
         buttons.setCellValueFactory(new PropertyValueFactory<>("DUMMY"));
 
         Callback<TableColumn<Author, Void>, TableCell<Author, Void>> cellFactory
@@ -106,7 +145,7 @@ public class AuthorScreen {
                     public TableCell call(final TableColumn<Author, Void> param) {
                         final TableCell<Author, Void> cell = new TableCell<Author, Void>() {
 
-                            final Button button123 = new Button("Just Do It");
+                            final Button button123 = new Button("Delete");
 
                             {
                                 button123.setOnAction((ActionEvent event) -> {
@@ -139,7 +178,49 @@ public class AuthorScreen {
 
     }
 
-    public HBox addSearchBar(Stage stage){
+    public void updateButton(){
+        TableColumn<Author, Void> buttons = new TableColumn("Update");
+        buttons.setCellValueFactory(new PropertyValueFactory<>("DUMMY"));
+
+        Callback<TableColumn<Author, Void>, TableCell<Author, Void>> cellFactory
+                = //
+                new Callback<TableColumn<Author, Void>, TableCell<Author, Void>>() {
+                    @Override
+                    public TableCell call(final TableColumn<Author, Void> param) {
+                        final TableCell<Author, Void> cell = new TableCell<Author, Void>() {
+
+                            final Button button123 = new Button("Update");
+
+                            {
+                                button123.setOnAction((ActionEvent event) -> {
+                                    Author author = authorTable.getSelectionModel().getSelectedItem();
+                                    AuthorRepository.updateByObject(author);
+                                });
+
+
+                            }
+
+                            @Override
+                            public void updateItem(Void item, boolean empty) {
+                                super.updateItem(item, empty);
+                                if (empty) {
+                                    setGraphic(null);
+
+                                } else {
+
+                                    setGraphic(button123);
+
+                                }
+                            }
+                        };
+                        return cell;
+                    }
+                };
+        buttons.setCellFactory(cellFactory);
+        authorTable.getColumns().add(buttons);
+    }
+
+    public HBox addSearchBar(Stage stage, boolean deleteButtonBool, boolean updateButtonBool){
 
         HBox root = new HBox();
         HBox choiseHBox = new HBox();
@@ -174,7 +255,7 @@ public class AuthorScreen {
                 authorsList = FXCollections.observableArrayList(AuthorRepository.getListByFirstName(searchField.getText()));
             }
             stackName = searchChoiceBox.getValue();
-            authorsScreen(stage);
+            authorsScreen(stage, deleteButtonBool, updateButtonBool);
         });
